调整项目结构，以及XML文档
