+++ .gitignore
@@ -259,7 +259,6 @@ paket-files/
 # Python Tools for Visual Studio (PTVS)
 __pycache__/
 *.pyc
-/Qincai.Api/Qincai.Api.xml
 
 /Qincai.Wx/project.config.json
 # Vant 组件
