+++ Qincai.Dtos/Config/QiniuConfig.cs
@@ -1,40 +1,6 @@
 ﻿namespace Qincai.Dtos
 {
     /// <summary>
-    /// JWT配置参数
-    /// </summary>
-    public class JwtConfig
-    {
-        /// <summary>
-        /// 密钥
-        /// </summary>
-        public string Secret { get; set; }
-        /// <summary>
-        /// 颁发者
-        /// </summary>
-        public string Issuer { get; set; }
-        /// <summary>
-        /// 接受者
-        /// </summary>
-        public string Audience { get; set; }
-    }
-
-    /// <summary>
-    /// 微信小程序配置参数
-    /// </summary>
-    public class WxOpenConfig
-    {
-        /// <summary>
-        /// 微信小程序 AppId
-        /// </summary>
-        public string AppId { get; set; }
-        /// <summary>
-        /// 微信小程序 Secert
-        /// </summary>
-        public string Secret { get; set; }
-    }
-
-    /// <summary>
     /// 七牛云配置参数
     /// </summary>
     public class QiniuConfig
