+++ Qincai.Dtos/Config/WxOpenConfig.cs
@@ -1,0 +1,17 @@
+﻿namespace Qincai.Dtos
+{
+    /// <summary>
+    /// 微信小程序配置参数
+    /// </summary>
+    public class WxOpenConfig
+    {
+        /// <summary>
+        /// 微信小程序 AppId
+        /// </summary>
+        public string AppId { get; set; }
+        /// <summary>
+        /// 微信小程序 Secert
+        /// </summary>
+        public string Secret { get; set; }
+    }
+}
