+++ Qincai.Dtos/Interface/IUserRegister.cs
@@ -1,0 +1,8 @@
+﻿namespace Qincai.Dtos
+{
+    public interface IUserRegister
+    {
+        string Name { get; set; }
+        string AvatarUrl { get; set; }
+    }
+}
