+++ Qincai.Dtos/PagedResult.cs
@@ -25,34 +25,4 @@ namespace Qincai.Dtos
         /// </summary>
         public int PageSize { get; set; }
     }
-
-    /// <summary>
-    /// 分页参数
-    /// </summary>
-    public interface IPagedParam
-    {
-        /// <summary>
-        /// 当前页数
-        /// </summary>
-        int Page { get; set; }
-        /// <summary>
-        /// 每页数量
-        /// </summary>
-        int PageSize { get; set; }
-    }
-
-    /// <summary>
-    /// 排序参数
-    /// </summary>
-    public interface ISortedParam
-    {
-        /// <summary>
-        /// 排序字段
-        /// </summary>
-        string OrderBy { get; set; }
-        /// <summary>
-        /// 是否降序
-        /// </summary>
-        bool Descending { get; set; }
-    }
 }
